RAG/12thclass_RAG/pyproject.toml
[project]
name = "12thclass-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb==0.6.3",
    "google-generativeai==0.7.2",
    "langchain[google-genai]==0.3.27",
    "langchain-community==0.3.28",
    "numpy==2.1.3",  # adjust to latest stable
    "openai==1.108.2",
    "pypdf==6.1.0",
    "python-dotenv==1.1.1",
    "tiktoken==0.11.0",
]
